requirements.txt
python-dateutil
requests
prettytable